Move GitTraversal into the project crate
