+++ crates/project/Cargo.toml
@@ -73,6 +73,7 @@ shlex.workspace = true
 smol.workspace = true
 snippet.workspace = true
 snippet_provider.workspace = true
+sum_tree.workspace = true
 task.workspace = true
 tempfile.workspace = true
 terminal.workspace = true
