+++ crates/worktree/src/worktree_tests.rs
@@ -1,15 +1,12 @@
 use crate::{
-    worktree_settings::WorktreeSettings, Entry, EntryKind, Event, PathChange, Snapshot,
-    WorkDirectory, Worktree, WorktreeModelHandle,
+    worktree_settings::WorktreeSettings, Entry, EntryKind, Event, PathChange, WorkDirectory,
+    Worktree, WorktreeModelHandle,
 };
 use anyhow::Result;
 use fs::{FakeFs, Fs, RealFs, RemoveOptions};
 use git::{
     repository::RepoPath,
-    status::{
-        FileStatus, GitSummary, StatusCode, TrackedStatus, TrackedSummary, UnmergedStatus,
-        UnmergedStatusCode,
-    },
+    status::{FileStatus, StatusCode, TrackedStatus},
     GITIGNORE,
 };
 use git2::RepositoryInitOptions;
@@ -27,7 +24,6 @@ use std::{
     mem,
     path::{Path, PathBuf},
     sync::Arc,
-    time::Duration,
 };
 use util::{path, test::TempTree, ResultExt};
 
@@ -1473,86 +1469,6 @@ async fn test_create_directory_during_initial_scan(cx: &mut TestAppContext) {
 }
 
 #[gpui::test]
-async fn test_bump_mtime_of_git_repo_workdir(cx: &mut TestAppContext) {
-    init_test(cx);
-
-    // Create a worktree with a git directory.
-    let fs = FakeFs::new(cx.background_executor.clone());
-    fs.insert_tree(
-        path!("/root"),
-        json!({
-            ".git": {},
-            "a.txt": "",
-            "b": {
-                "c.txt": "",
-            },
-        }),
-    )
-    .await;
-    fs.set_head_and_index_for_repo(
-        path!("/root/.git").as_ref(),
-        &[("a.txt".into(), "".into()), ("b/c.txt".into(), "".into())],
-    );
-    cx.run_until_parked();
-
-    let tree = Worktree::local(
-        path!("/root").as_ref(),
-        true,
-        fs.clone(),
-        Default::default(),
-        &mut cx.to_async(),
-    )
-    .await
-    .unwrap();
-    cx.executor().run_until_parked();
-
-    let (old_entry_ids, old_mtimes) = tree.read_with(cx, |tree, _| {
-        (
-            tree.entries(true, 0).map(|e| e.id).collect::<Vec<_>>(),
-            tree.entries(true, 0).map(|e| e.mtime).collect::<Vec<_>>(),
-        )
-    });
-
-    // Regression test: after the directory is scanned, touch the git repo's
-    // working directory, bumping its mtime. That directory keeps its project
-    // entry id after the directories are re-scanned.
-    fs.touch_path(path!("/root")).await;
-    cx.executor().run_until_parked();
-
-    let (new_entry_ids, new_mtimes) = tree.read_with(cx, |tree, _| {
-        (
-            tree.entries(true, 0).map(|e| e.id).collect::<Vec<_>>(),
-            tree.entries(true, 0).map(|e| e.mtime).collect::<Vec<_>>(),
-        )
-    });
-    assert_eq!(new_entry_ids, old_entry_ids);
-    assert_ne!(new_mtimes, old_mtimes);
-
-    // Regression test: changes to the git repository should still be
-    // detected.
-    fs.set_head_for_repo(
-        path!("/root/.git").as_ref(),
-        &[
-            ("a.txt".into(), "".into()),
-            ("b/c.txt".into(), "something-else".into()),
-        ],
-    );
-    cx.executor().run_until_parked();
-    cx.executor().advance_clock(Duration::from_secs(1));
-
-    let snapshot = tree.read_with(cx, |tree, _| tree.snapshot());
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new(""), MODIFIED),
-            (Path::new("a.txt"), GitSummary::UNCHANGED),
-            (Path::new("b/c.txt"), MODIFIED),
-        ],
-    );
-}
-
-#[gpui::test]
 async fn test_create_dir_all_on_create_entry(cx: &mut TestAppContext) {
     init_test(cx);
     cx.executor().allow_parking();
@@ -2195,11 +2111,6 @@ fn random_filename(rng: &mut impl Rng) -> String {
         .map(char::from)
         .collect()
 }
-
-const CONFLICT: FileStatus = FileStatus::Unmerged(UnmergedStatus {
-    first_head: UnmergedStatusCode::Updated,
-    second_head: UnmergedStatusCode::Updated,
-});
 
 // NOTE:
 // This test always fails on Windows, because on Windows, unlike on Unix, you can't rename
@@ -2921,435 +2832,6 @@ async fn test_repository_subfolder_git_status(cx: &mut TestAppContext) {
 }
 
 #[gpui::test]
-async fn test_traverse_with_git_status(cx: &mut TestAppContext) {
-    init_test(cx);
-    let fs = FakeFs::new(cx.background_executor.clone());
-    fs.insert_tree(
-        path!("/root"),
-        json!({
-            "x": {
-                ".git": {},
-                "x1.txt": "foo",
-                "x2.txt": "bar",
-                "y": {
-                    ".git": {},
-                    "y1.txt": "baz",
-                    "y2.txt": "qux"
-                },
-                "z.txt": "sneaky..."
-            },
-            "z": {
-                ".git": {},
-                "z1.txt": "quux",
-                "z2.txt": "quuux"
-            }
-        }),
-    )
-    .await;
-
-    fs.set_status_for_repo(
-        Path::new(path!("/root/x/.git")),
-        &[
-            (Path::new("x2.txt"), StatusCode::Modified.index()),
-            (Path::new("z.txt"), StatusCode::Added.index()),
-        ],
-    );
-    fs.set_status_for_repo(
-        Path::new(path!("/root/x/y/.git")),
-        &[(Path::new("y1.txt"), CONFLICT)],
-    );
-    fs.set_status_for_repo(
-        Path::new(path!("/root/z/.git")),
-        &[(Path::new("z2.txt"), StatusCode::Added.index())],
-    );
-
-    let tree = Worktree::local(
-        Path::new(path!("/root")),
-        true,
-        fs.clone(),
-        Default::default(),
-        &mut cx.to_async(),
-    )
-    .await
-    .unwrap();
-
-    tree.flush_fs_events(cx).await;
-    cx.read(|cx| tree.read(cx).as_local().unwrap().scan_complete())
-        .await;
-    cx.executor().run_until_parked();
-
-    let snapshot = tree.read_with(cx, |tree, _| tree.snapshot());
-
-    let mut traversal = snapshot
-        .traverse_from_path(true, false, true, Path::new("x"))
-        .with_git_statuses();
-
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("x/x1.txt"));
-    assert_eq!(entry.git_summary, GitSummary::UNCHANGED);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("x/x2.txt"));
-    assert_eq!(entry.git_summary, MODIFIED);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("x/y/y1.txt"));
-    assert_eq!(entry.git_summary, GitSummary::CONFLICT);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("x/y/y2.txt"));
-    assert_eq!(entry.git_summary, GitSummary::UNCHANGED);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("x/z.txt"));
-    assert_eq!(entry.git_summary, ADDED);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("z/z1.txt"));
-    assert_eq!(entry.git_summary, GitSummary::UNCHANGED);
-    let entry = traversal.next().unwrap();
-    assert_eq!(entry.path.as_ref(), Path::new("z/z2.txt"));
-    assert_eq!(entry.git_summary, ADDED);
-}
-
-#[gpui::test]
-async fn test_propagate_git_statuses(cx: &mut TestAppContext) {
-    init_test(cx);
-    let fs = FakeFs::new(cx.background_executor.clone());
-    fs.insert_tree(
-        path!("/root"),
-        json!({
-            ".git": {},
-            "a": {
-                "b": {
-                    "c1.txt": "",
-                    "c2.txt": "",
-                },
-                "d": {
-                    "e1.txt": "",
-                    "e2.txt": "",
-                    "e3.txt": "",
-                }
-            },
-            "f": {
-                "no-status.txt": ""
-            },
-            "g": {
-                "h1.txt": "",
-                "h2.txt": ""
-            },
-        }),
-    )
-    .await;
-
-    fs.set_status_for_repo(
-        Path::new(path!("/root/.git")),
-        &[
-            (Path::new("a/b/c1.txt"), StatusCode::Added.index()),
-            (Path::new("a/d/e2.txt"), StatusCode::Modified.index()),
-            (Path::new("g/h2.txt"), CONFLICT),
-        ],
-    );
-
-    let tree = Worktree::local(
-        Path::new(path!("/root")),
-        true,
-        fs.clone(),
-        Default::default(),
-        &mut cx.to_async(),
-    )
-    .await
-    .unwrap();
-
-    cx.read(|cx| tree.read(cx).as_local().unwrap().scan_complete())
-        .await;
-
-    cx.executor().run_until_parked();
-    let snapshot = tree.read_with(cx, |tree, _| tree.snapshot());
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new(""), GitSummary::CONFLICT + MODIFIED + ADDED),
-            (Path::new("g"), GitSummary::CONFLICT),
-            (Path::new("g/h2.txt"), GitSummary::CONFLICT),
-        ],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new(""), GitSummary::CONFLICT + ADDED + MODIFIED),
-            (Path::new("a"), ADDED + MODIFIED),
-            (Path::new("a/b"), ADDED),
-            (Path::new("a/b/c1.txt"), ADDED),
-            (Path::new("a/b/c2.txt"), GitSummary::UNCHANGED),
-            (Path::new("a/d"), MODIFIED),
-            (Path::new("a/d/e2.txt"), MODIFIED),
-            (Path::new("f"), GitSummary::UNCHANGED),
-            (Path::new("f/no-status.txt"), GitSummary::UNCHANGED),
-            (Path::new("g"), GitSummary::CONFLICT),
-            (Path::new("g/h2.txt"), GitSummary::CONFLICT),
-        ],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("a/b"), ADDED),
-            (Path::new("a/b/c1.txt"), ADDED),
-            (Path::new("a/b/c2.txt"), GitSummary::UNCHANGED),
-            (Path::new("a/d"), MODIFIED),
-            (Path::new("a/d/e1.txt"), GitSummary::UNCHANGED),
-            (Path::new("a/d/e2.txt"), MODIFIED),
-            (Path::new("f"), GitSummary::UNCHANGED),
-            (Path::new("f/no-status.txt"), GitSummary::UNCHANGED),
-            (Path::new("g"), GitSummary::CONFLICT),
-        ],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("a/b/c1.txt"), ADDED),
-            (Path::new("a/b/c2.txt"), GitSummary::UNCHANGED),
-            (Path::new("a/d/e1.txt"), GitSummary::UNCHANGED),
-            (Path::new("a/d/e2.txt"), MODIFIED),
-            (Path::new("f/no-status.txt"), GitSummary::UNCHANGED),
-        ],
-    );
-}
-
-#[gpui::test]
-async fn test_propagate_statuses_for_repos_under_project(cx: &mut TestAppContext) {
-    init_test(cx);
-    let fs = FakeFs::new(cx.background_executor.clone());
-    fs.insert_tree(
-        path!("/root"),
-        json!({
-            "x": {
-                ".git": {},
-                "x1.txt": "foo",
-                "x2.txt": "bar"
-            },
-            "y": {
-                ".git": {},
-                "y1.txt": "baz",
-                "y2.txt": "qux"
-            },
-            "z": {
-                ".git": {},
-                "z1.txt": "quux",
-                "z2.txt": "quuux"
-            }
-        }),
-    )
-    .await;
-
-    fs.set_status_for_repo(
-        Path::new(path!("/root/x/.git")),
-        &[(Path::new("x1.txt"), StatusCode::Added.index())],
-    );
-    fs.set_status_for_repo(
-        Path::new(path!("/root/y/.git")),
-        &[
-            (Path::new("y1.txt"), CONFLICT),
-            (Path::new("y2.txt"), StatusCode::Modified.index()),
-        ],
-    );
-    fs.set_status_for_repo(
-        Path::new(path!("/root/z/.git")),
-        &[(Path::new("z2.txt"), StatusCode::Modified.index())],
-    );
-
-    let tree = Worktree::local(
-        Path::new(path!("/root")),
-        true,
-        fs.clone(),
-        Default::default(),
-        &mut cx.to_async(),
-    )
-    .await
-    .unwrap();
-
-    tree.flush_fs_events(cx).await;
-    cx.read(|cx| tree.read(cx).as_local().unwrap().scan_complete())
-        .await;
-    cx.executor().run_until_parked();
-
-    let snapshot = tree.read_with(cx, |tree, _| tree.snapshot());
-
-    check_git_statuses(
-        &snapshot,
-        &[(Path::new("x"), ADDED), (Path::new("x/x1.txt"), ADDED)],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("y"), GitSummary::CONFLICT + MODIFIED),
-            (Path::new("y/y1.txt"), GitSummary::CONFLICT),
-            (Path::new("y/y2.txt"), MODIFIED),
-        ],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("z"), MODIFIED),
-            (Path::new("z/z2.txt"), MODIFIED),
-        ],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[(Path::new("x"), ADDED), (Path::new("x/x1.txt"), ADDED)],
-    );
-
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("x"), ADDED),
-            (Path::new("x/x1.txt"), ADDED),
-            (Path::new("x/x2.txt"), GitSummary::UNCHANGED),
-            (Path::new("y"), GitSummary::CONFLICT + MODIFIED),
-            (Path::new("y/y1.txt"), GitSummary::CONFLICT),
-            (Path::new("y/y2.txt"), MODIFIED),
-            (Path::new("z"), MODIFIED),
-            (Path::new("z/z1.txt"), GitSummary::UNCHANGED),
-            (Path::new("z/z2.txt"), MODIFIED),
-        ],
-    );
-}
-
-#[gpui::test]
-async fn test_propagate_statuses_for_nested_repos(cx: &mut TestAppContext) {
-    init_test(cx);
-    let fs = FakeFs::new(cx.background_executor.clone());
-    fs.insert_tree(
-        path!("/root"),
-        json!({
-            "x": {
-                ".git": {},
-                "x1.txt": "foo",
-                "x2.txt": "bar",
-                "y": {
-                    ".git": {},
-                    "y1.txt": "baz",
-                    "y2.txt": "qux"
-                },
-                "z.txt": "sneaky..."
-            },
-            "z": {
-                ".git": {},
-                "z1.txt": "quux",
-                "z2.txt": "quuux"
-            }
-        }),
-    )
-    .await;
-
-    fs.set_status_for_repo(
-        Path::new(path!("/root/x/.git")),
-        &[
-            (Path::new("x2.txt"), StatusCode::Modified.index()),
-            (Path::new("z.txt"), StatusCode::Added.index()),
-        ],
-    );
-    fs.set_status_for_repo(
-        Path::new(path!("/root/x/y/.git")),
-        &[(Path::new("y1.txt"), CONFLICT)],
-    );
-
-    fs.set_status_for_repo(
-        Path::new(path!("/root/z/.git")),
-        &[(Path::new("z2.txt"), StatusCode::Added.index())],
-    );
-
-    let tree = Worktree::local(
-        Path::new(path!("/root")),
-        true,
-        fs.clone(),
-        Default::default(),
-        &mut cx.to_async(),
-    )
-    .await
-    .unwrap();
-
-    tree.flush_fs_events(cx).await;
-    cx.read(|cx| tree.read(cx).as_local().unwrap().scan_complete())
-        .await;
-    cx.executor().run_until_parked();
-
-    let snapshot = tree.read_with(cx, |tree, _| tree.snapshot());
-
-    // Sanity check the propagation for x/y and z
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("x/y"), GitSummary::CONFLICT),
-            (Path::new("x/y/y1.txt"), GitSummary::CONFLICT),
-            (Path::new("x/y/y2.txt"), GitSummary::UNCHANGED),
-        ],
-    );
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("z"), ADDED),
-            (Path::new("z/z1.txt"), GitSummary::UNCHANGED),
-            (Path::new("z/z2.txt"), ADDED),
-        ],
-    );
-
-    // Test one of the fundamental cases of propagation blocking, the transition from one git repository to another
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("x"), MODIFIED + ADDED),
-            (Path::new("x/y"), GitSummary::CONFLICT),
-            (Path::new("x/y/y1.txt"), GitSummary::CONFLICT),
-        ],
-    );
-
-    // Sanity check everything around it
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new("x"), MODIFIED + ADDED),
-            (Path::new("x/x1.txt"), GitSummary::UNCHANGED),
-            (Path::new("x/x2.txt"), MODIFIED),
-            (Path::new("x/y"), GitSummary::CONFLICT),
-            (Path::new("x/y/y1.txt"), GitSummary::CONFLICT),
-            (Path::new("x/y/y2.txt"), GitSummary::UNCHANGED),
-            (Path::new("x/z.txt"), ADDED),
-        ],
-    );
-
-    // Test the other fundamental case, transitioning from git repository to non-git repository
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new(""), GitSummary::UNCHANGED),
-            (Path::new("x"), MODIFIED + ADDED),
-            (Path::new("x/x1.txt"), GitSummary::UNCHANGED),
-        ],
-    );
-
-    // And all together now
-    check_git_statuses(
-        &snapshot,
-        &[
-            (Path::new(""), GitSummary::UNCHANGED),
-            (Path::new("x"), MODIFIED + ADDED),
-            (Path::new("x/x1.txt"), GitSummary::UNCHANGED),
-            (Path::new("x/x2.txt"), MODIFIED),
-            (Path::new("x/y"), GitSummary::CONFLICT),
-            (Path::new("x/y/y1.txt"), GitSummary::CONFLICT),
-            (Path::new("x/y/y2.txt"), GitSummary::UNCHANGED),
-            (Path::new("x/z.txt"), ADDED),
-            (Path::new("z"), ADDED),
-            (Path::new("z/z1.txt"), GitSummary::UNCHANGED),
-            (Path::new("z/z2.txt"), ADDED),
-        ],
-    );
-}
-
-#[gpui::test]
 async fn test_conflicted_cherry_pick(cx: &mut TestAppContext) {
     init_test(cx);
     cx.executor().allow_parking();
@@ -3489,34 +2971,6 @@ fn test_unrelativize() {
         Path::new("../../README.md").into()
     );
 }
-
-#[track_caller]
-fn check_git_statuses(snapshot: &Snapshot, expected_statuses: &[(&Path, GitSummary)]) {
-    let mut traversal = snapshot
-        .traverse_from_path(true, true, false, "".as_ref())
-        .with_git_statuses();
-    let found_statuses = expected_statuses
-        .iter()
-        .map(|&(path, _)| {
-            let git_entry = traversal
-                .find(|git_entry| &*git_entry.path == path)
-                .unwrap_or_else(|| panic!("Traversal has no entry for {path:?}"));
-            (path, git_entry.git_summary)
-        })
-        .collect::<Vec<_>>();
-    assert_eq!(found_statuses, expected_statuses);
-}
-
-const ADDED: GitSummary = GitSummary {
-    index: TrackedSummary::ADDED,
-    count: 1,
-    ..GitSummary::UNCHANGED
-};
-const MODIFIED: GitSummary = GitSummary {
-    index: TrackedSummary::MODIFIED,
-    count: 1,
-    ..GitSummary::UNCHANGED
-};
 
 #[track_caller]
 fn git_init(path: &Path) -> git2::Repository {
